+++ stable_baselines3/common/atari_wrappers.py
@@ -1,31 +1,213 @@
 import gym
-from gym.wrappers import AtariPreprocessing
+from gym import spaces
 import numpy as np
+try:
+    import cv2  # pytype:disable=import-error
+    cv2.ocl.setUseOpenCL(False)
+except ImportError:
+    cv2 = None
 
 from stable_baselines3.common.type_aliases import GymStepReturn
+
+
+class NoopResetEnv(gym.Wrapper):
+    def __init__(self, env: gym.Env, noop_max: int = 30):
+        """
+        Sample initial states by taking random number of no-ops on reset.
+        No-op is assumed to be action 0.
+
+        :param env: (gym.Env) the environment to wrap
+        :param noop_max: (int) the maximum value of no-ops to run
+        """
+        gym.Wrapper.__init__(self, env)
+        self.noop_max = noop_max
+        self.override_num_noops = None
+        self.noop_action = 0
+        assert env.unwrapped.get_action_meanings()[0] == 'NOOP'
+
+    def reset(self, **kwargs) -> np.ndarray:
+        self.env.reset(**kwargs)
+        if self.override_num_noops is not None:
+            noops = self.override_num_noops
+        else:
+            noops = self.unwrapped.np_random.randint(1, self.noop_max + 1)
+        assert noops > 0
+        obs = np.empty(0)
+        for _ in range(noops):
+            obs, _, done, _ = self.env.step(self.noop_action)
+            if done:
+                obs = self.env.reset(**kwargs)
+        return obs
+
+
+class FireResetEnv(gym.Wrapper):
+    def __init__(self, env: gym.Env):
+        """
+        Take action on reset for environments that are fixed until firing.
+
+        :param env: (gym.Env) the environment to wrap
+        """
+        gym.Wrapper.__init__(self, env)
+        assert env.unwrapped.get_action_meanings()[1] == 'FIRE'
+        assert len(env.unwrapped.get_action_meanings()) >= 3
+
+    def reset(self, **kwargs) -> np.ndarray:
+        self.env.reset(**kwargs)
+        obs, _, done, _ = self.env.step(1)
+        if done:
+            self.env.reset(**kwargs)
+        obs, _, done, _ = self.env.step(2)
+        if done:
+            self.env.reset(**kwargs)
+        return obs
+
+
+class EpisodicLifeEnv(gym.Wrapper):
+    def __init__(self, env: gym.Env):
+        """
+        Make end-of-life == end-of-episode, but only reset on true game over.
+        Done by DeepMind for the DQN and co. since it helps value estimation.
+
+        :param env: (gym.Env) the environment to wrap
+        """
+        gym.Wrapper.__init__(self, env)
+        self.lives = 0
+        self.was_real_done = True
+
+    def step(self, action: int) -> GymStepReturn:
+        obs, reward, done, info = self.env.step(action)
+        self.was_real_done = done
+        # check current lives, make loss of life terminal,
+        # then update lives to handle bonus lives
+        lives = self.env.unwrapped.ale.lives()
+        if 0 < lives < self.lives:
+            # for Qbert sometimes we stay in lives == 0 condtion for a few frames
+            # so its important to keep lives > 0, so that we only reset once
+            # the environment advertises done.
+            done = True
+        self.lives = lives
+        return obs, reward, done, info
+
+    def reset(self, **kwargs) -> np.ndarray:
+        """
+        Calls the Gym environment reset, only when lives are exhausted.
+        This way all states are still reachable even though lives are episodic,
+        and the learner need not know about any of this behind-the-scenes.
+
+        :param kwargs: Extra keywords passed to env.reset() call
+        :return: (np.ndarray) the first observation of the environment
+        """
+        if self.was_real_done:
+            obs = self.env.reset(**kwargs)
+        else:
+            # no-op step to advance from terminal/lost life state
+            obs, _, _, _ = self.env.step(0)
+        self.lives = self.env.unwrapped.ale.lives()
+        return obs
+
+
+class MaxAndSkipEnv(gym.Wrapper):
+    def __init__(self, env: gym.Env, skip: int = 4):
+        """
+        Return only every ``skip``-th frame (frameskipping)
+
+        :param env: (gym.Env) the environment
+        :param skip: (int) number of ``skip``-th frame
+        """
+        gym.Wrapper.__init__(self, env)
+        # most recent raw observations (for max pooling across time steps)
+        self._obs_buffer = np.zeros((2,) + env.observation_space.shape, dtype=env.observation_space.dtype)
+        self._skip = skip
+
+    def step(self, action: int) -> GymStepReturn:
+        """
+        Step the environment with the given action
+        Repeat action, sum reward, and max over last observations.
+
+        :param action: ([int] or [float]) the action
+        :return: ([int] or [float], [float], [bool], dict) observation, reward, done, information
+        """
+        total_reward = 0.0
+        done = None
+        for i in range(self._skip):
+            obs, reward, done, info = self.env.step(action)
+            if i == self._skip - 2:
+                self._obs_buffer[0] = obs
+            if i == self._skip - 1:
+                self._obs_buffer[1] = obs
+            total_reward += reward
+            if done:
+                break
+        # Note that the observation on the done=True frame
+        # doesn't matter
+        max_frame = self._obs_buffer.max(axis=0)
+
+        return max_frame, total_reward, done, info
+
+    def reset(self, **kwargs):
+        return self.env.reset(**kwargs)
+
+
+class ClipRewardEnv(gym.RewardWrapper):
+    def __init__(self, env: gym.Env):
+        """
+        Clips the reward to {+1, 0, -1} by its sign.
+
+        :param env: (gym.Env) the environment
+        """
+        gym.RewardWrapper.__init__(self, env)
+
+    def reward(self, reward: float) -> float:
+        """
+        Bin reward to {+1, 0, -1} by its sign.
+
+        :param reward: (float)
+        :return: (float)
+        """
+        return np.sign(reward)
+
+
+class WarpFrame(gym.ObservationWrapper):
+    def __init__(self, env: gym.Env, width: int = 84, height: int = 84):
+        """
+        Convert to grayscale and warp frames to 84x84 (default)
+        as done in the Nature paper and later work.
+
+        :param env: (gym.Env) the environment
+        :param width: (int)
+        :param height: (int)
+        """
+        gym.ObservationWrapper.__init__(self, env)
+        self.width = width
+        self.height = height
+        self.observation_space = spaces.Box(low=0, high=255, shape=(self.height, self.width, 1),
+                                            dtype=env.observation_space.dtype)
+
+    def observation(self, frame: np.ndarray) -> np.ndarray:
+        """
+        returns the current observation from a frame
+
+        :param frame: (np.ndarray) environment frame
+        :return: (np.ndarray) the observation
+        """
+        frame = cv2.cvtColor(frame, cv2.COLOR_RGB2GRAY)
+        frame = cv2.resize(frame, (self.width, self.height), interpolation=cv2.INTER_AREA)
+        return frame[:, :, None]
 
 
 class AtariWrapper(gym.Wrapper):
     """
     Atari 2600 preprocessings
 
-    It is a wrapper around the one found in gym.
-    It reshapes the observation to have an additional dimension and clip the reward.
-    See https://github.com/openai/gym/blob/master/gym/wrappers/atari_preprocessing.py
-    .
-    This class follows the guidelines in
-    Machado et al. (2018), "Revisiting the Arcade Learning Environment:
-    Evaluation Protocols and Open Problems for General Agents".
-
     Specifically:
 
     * NoopReset: obtain initial state by taking random number of no-ops on reset.
     * Frame skipping: 4 by default
     * Max-pooling: most recent two observations
-    * Termination signal when a life is lost: turned off by default. Not recommended by Machado et al. (2018).
+    * Termination signal when a life is lost.
     * Resize to a square image: 84x84 by default
-    * Grayscale observation: by default
-    * Scale observation: optional
+    * Grayscale observation
+    * Clip reward to {-1, 0, 1}
 
     :param env: (gym.Env) gym environment
     :param noop_max: (int): max number of no-ops
@@ -33,44 +215,22 @@ class AtariWrapper(gym.Wrapper):
     :param screen_size: (int): resize Atari frame
     :param terminal_on_life_loss: (bool): if True, then step() returns done=True whenever a
             life is lost.
-    :param grayscale_obs: (bool): if True (default), then gray scale observation is returned, otherwise, RGB observation
-            is returned.
-    :param scale_obs: (bool): if True, then observation normalized in range [0,1] is returned. It also limits memory
-            optimization benefits of FrameStack Wrapper.
-    :param scale_obs: (bool) If True (default), the reward is clip to {-1, 0, 1} depending on its sign.
+    :param clip_reward: (bool) If True (default), the reward is clip to {-1, 0, 1} depending on its sign.
     """
     def __init__(self, env: gym.Env,
                  noop_max: int = 30,
                  frame_skip: int = 4,
                  screen_size: int = 84,
-                 terminal_on_life_loss: bool = False,
-                 grayscale_obs: bool = True,
-                 scale_obs: bool = False,
+                 terminal_on_life_loss: bool = True,
                  clip_reward: bool = True):
-        env = AtariPreprocessing(env, noop_max=noop_max, frame_skip=frame_skip, screen_size=screen_size,
-                                 terminal_on_life_loss=terminal_on_life_loss, grayscale_obs=grayscale_obs,
-                                 scale_obs=scale_obs)
-        # Add channel dimension
-        if grayscale_obs:
-            obs_space = env.observation_space
-            _low, _high, _obs_dtype = (0, 255, np.uint8) if not scale_obs else (0, 1, np.float32)
-            env.observation_space = gym.spaces.Box(low=_low, high=_high, shape=obs_space.shape + (1,),
-                                                   dtype=_obs_dtype)
+        env = NoopResetEnv(env, noop_max=noop_max)
+        env = MaxAndSkipEnv(env, skip=frame_skip)
+        if terminal_on_life_loss:
+            env = EpisodicLifeEnv(env)
+        if 'FIRE' in env.unwrapped.get_action_meanings():
+            env = FireResetEnv(env)
+        env = WarpFrame(env, width=screen_size, height=screen_size)
+        if clip_reward:
+            env = ClipRewardEnv(env)
 
         super(AtariWrapper, self).__init__(env)
-        self.clip_reward = clip_reward
-
-    def _add_axis(self, obs: np.ndarray) -> np.ndarray:
-        if self.env.grayscale_obs:
-            return obs[..., np.newaxis]
-        return obs
-
-    def reset(self) -> np.ndarray:
-        return self._add_axis(self.env.reset())
-
-    def step(self, action: int) -> GymStepReturn:
-        obs, reward, done, info = self.env.step(action)
-        # Bin reward to {+1, 0, -1} by its sign.
-        if self.clip_reward:
-            reward = np.sign(reward)
-        return self._add_axis(obs), reward, done, info