+++ tests/test_callbacks.py
@@ -4,21 +4,24 @@ import shutil
 import pytest
 import gym
 
-from stable_baselines3 import A2C, PPO, SAC, TD3
+from stable_baselines3 import A2C, PPO, SAC, TD3, DQN
 from stable_baselines3.common.callbacks import (CallbackList, CheckpointCallback, EvalCallback,
                                                 EveryNTimesteps, StopTrainingOnRewardThreshold)
 
 
-@pytest.mark.parametrize("model_class", [A2C, PPO, SAC, TD3])
-def test_callbacks(model_class):
-    log_folder = './logs/callbacks/'
+@pytest.mark.parametrize("model_class", [A2C, PPO, SAC, TD3, DQN])
+def test_callbacks(tmp_path, model_class):
+    log_folder = tmp_path / 'logs/callbacks/'
+
+    # Dyn only support discrete actions
+    env_name = select_env(model_class)
     # Create RL model
     # Small network for fast test
-    model = model_class('MlpPolicy', 'Pendulum-v0', policy_kwargs=dict(net_arch=[32]))
+    model = model_class('MlpPolicy', env_name, policy_kwargs=dict(net_arch=[32]))
 
     checkpoint_callback = CheckpointCallback(save_freq=1000, save_path=log_folder)
 
-    eval_env = gym.make('Pendulum-v0')
+    eval_env = gym.make(env_name)
     # Stop training if the performance is good enough
     callback_on_best = StopTrainingOnRewardThreshold(reward_threshold=-1200, verbose=1)
 
@@ -42,3 +45,10 @@ def test_callbacks(model_class):
     model.learn(500, callback=lambda _locals, _globals: True)
     if os.path.exists(log_folder):
         shutil.rmtree(log_folder)
+
+
+def select_env(model_class) -> str:
+    if model_class is DQN:
+        return 'CartPole-v0'
+    else:
+        return 'Pendulum-v0'